automatically set parent as walk down operation tree
