+++ src/Compilers/Core/Portable/Operations/OperationExtensions.cs
@@ -2,11 +2,9 @@
 
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using Microsoft.CodeAnalysis.PooledObjects;
-using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Semantics
@@ -95,111 +93,6 @@ namespace Microsoft.CodeAnalysis.Semantics
             }
 
             return arrayBuilder.ToImmutableAndFree();
-        }
-
-        private static IOperation WalkDownOperationToFindParent(
-            HashSet<IOperation> operationAlreadyProcessed, IOperation operation, TextSpan span)
-        {
-            void EnqueueChildOperations(Queue<IOperation> queue, IOperation parent)
-            {
-                foreach (var o in parent.Children)
-                {
-                    queue.Enqueue(o);
-                }
-            }
-
-            // do we have a pool for queue?
-            var parentChildQueue = new Queue<IOperation>();
-            EnqueueChildOperations(parentChildQueue, operation);
-
-            // walk down the child operation to find parent operation
-            // every child returned by the queue should already have Parent operation set
-            IOperation child;
-            while ((child = parentChildQueue.Dequeue()) != null)
-            {
-                if (!operationAlreadyProcessed.Add(child))
-                {
-                    // don't process IOperation we already processed otherwise,
-                    // we can walk down same tree multiple times
-                    continue;
-                }
-
-                if (child == operation)
-                {
-                    // parent found
-                    return child.Parent;
-                }
-
-                if (!child.Syntax.FullSpan.IntersectsWith(span))
-                {
-                    // not related node, don't walk down
-                    continue;
-                }
-
-                // queue children so that we can do breadth first search
-                EnqueueChildOperations(parentChildQueue, child);
-            }
-
-            return null;
-        }
-
-        internal static IOperation FindParentOperation(this SemanticModel semanticModel, IOperation operation)
-        {
-            Debug.Assert(operation != null);
-
-            // do we have a pool for hashset?
-            var operationAlreadyProcessed = new HashSet<IOperation>();
-
-            var targetNode = operation.Syntax;
-            var currentCandidate = targetNode.Parent;
-
-            while (currentCandidate != null)
-            {
-                Debug.Assert(currentCandidate.FullSpan.IntersectsWith(targetNode.FullSpan));
-
-                foreach (var childNode in currentCandidate.ChildNodes())
-                {
-                    if (!childNode.FullSpan.IntersectsWith(targetNode.FullSpan))
-                    {
-                        // skip unrelated node
-                        continue;
-                    }
-
-                    // get child operation
-                    var childOperation = semanticModel.GetOperationInternal(childNode);
-                    if (childOperation != null)
-                    {
-                        // there is no operation for this node
-                        continue;
-                    }
-
-                    // record we have processed this node
-                    if (!operationAlreadyProcessed.Add(childOperation))
-                    {
-                        // we already processed this tree. no need to dig down
-                        continue;
-                    }
-
-                    // check easy case first
-                    if (childOperation == operation)
-                    {
-                        // found parent, go up the spine until we found non-null parent Operation
-                        return currentCandidate.AncestorsAndSelf().Select(n => semanticModel.GetOperationInternal(n)).WhereNotNull().FirstOrDefault();
-                    }
-
-                    // walk down child operation tree to see whether sub tree contains the given operation
-                    var parent = WalkDownOperationToFindParent(operationAlreadyProcessed, childOperation, targetNode.FullSpan);
-                    if (parent != null)
-                    {
-                        return parent;
-                    }
-                }
-
-                currentCandidate = currentCandidate.Parent;
-            }
-
-            // root node. there is no parent
-            return null;
         }
     }
